requirements.dev.txt
aiohttp==3.8.5 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
annotated-types==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
black==23.9.1 ; python_version >= "3.9" and python_version < "4.0"
cfgv==3.4.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
distlib==0.3.7 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.1.3 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.103.2 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.12.4 ; python_version >= "3.9" and python_version < "4.0"
flake8==6.1.0 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.4.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httptools==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
identify==2.5.30 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
isort==5.12.0 ; python_version >= "3.9" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.9" and python_version < "4.0"
pathspec==0.11.2 ; python_version >= "3.9" and python_version < "4.0"
platformdirs==3.11.0 ; python_version >= "3.9" and python_version < "4.0"
pre-commit==3.4.0 ; python_version >= "3.9" and python_version < "4.0"
pycodestyle==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.10.1 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.4.2 ; python_version >= "3.9" and python_version < "4.0"
pyflakes==3.1.0 ; python_version >= "3.9" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.9" and python_version < "4.0"
slack-sdk==3.22.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
uvicorn[standard]==0.23.2 ; python_version >= "3.9" and python_version < "4.0"
uvloop==0.17.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_version < "4.0"
virtualenv==20.24.5 ; python_version >= "3.9" and python_version < "4.0"
watchfiles==0.20.0 ; python_version >= "3.9" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.9" and python_version < "4.0"
